+++ lib/registerPlugins.js
@@ -18,7 +18,6 @@ async function registerDefaultPlugins(server) {
         '../plugins/logging',
         '../plugins/swagger',
         '../plugins/template-validator',
-        '../plugins/pipeline-template-validator',
         '../plugins/command-validator',
         '../plugins/promster',
         '../plugins/metrics',
