+++ package.json
@@ -112,7 +112,7 @@
     "screwdriver-executor-k8s": "^15.0.1",
     "screwdriver-executor-k8s-vm": "^4.3.3",
     "screwdriver-executor-queue": "^4.0.0",
-    "screwdriver-executor-router": "^3.0.0", 
+    "screwdriver-executor-router": "^3.0.0",
     "screwdriver-logger": "^2.0.0",
     "screwdriver-models": "^29.11.0",
     "screwdriver-notifications-email": "^3.0.0",
