+++ test/lib/registerPlugins.test.js
@@ -17,7 +17,6 @@ describe('Register Unit Test Case', () => {
         '../../plugins/logging',
         '../../plugins/swagger',
         '../../plugins/template-validator',
-        '../../plugins/pipeline-template-validator',
         '../../plugins/command-validator',
         '../../plugins/promster',
         '../../plugins/metrics',
