+++ test/plugins/pipelines.test.js
@@ -4005,4 +4005,102 @@ describe('pipeline plugin test', () => {
             });
         });
     });
+
+    describe('POST /pipeline/template/validate', () => {
+        it('returns OK for a successful template yaml', () =>
+            server
+                .inject({
+                    method: 'POST',
+                    url: '/pipeline/template/validate',
+                    payload: TEMPLATE_VALID
+                })
+                .then(reply => {
+                    assert.strictEqual(reply.statusCode, 200);
+
+                    const payload = JSON.parse(reply.payload);
+
+                    assert.deepEqual(payload, {
+                        errors: [],
+                        template: {
+                            namespace: 'template_namespace',
+                            name: 'template_name',
+                            version: '1.2.3',
+                            description: 'template description',
+                            maintainer: 'name@domain.org',
+                            config: {
+                                jobs: { main: { steps: [{ init: 'npm install' }, { test: 'npm test' }] } },
+                                shared: {},
+                                parameters: {}
+                            }
+                        }
+                    });
+                }));
+
+        it('returns OK and error yaml for bad yaml', () =>
+            server
+                .inject({
+                    method: 'POST',
+                    url: '/pipeline/template/validate',
+                    payload: TEMPLATE_INVALID
+                })
+                .then(reply => {
+                    assert.strictEqual(reply.statusCode, 200);
+
+                    const payload = JSON.parse(reply.payload);
+
+                    assert.deepEqual(payload.template, {
+                        namespace: 'template_namespace',
+                        name: 'template_name',
+                        description: 'template description',
+                        maintainer: 'name@domain.org',
+                        config: {
+                            jobs: { main: { steps: [{ init: 'npm install' }, { test: 'npm test' }] } }
+                        }
+                    });
+
+                    assert.deepEqual(payload.errors, [
+                        {
+                            context: {
+                                key: 'version',
+                                label: 'version'
+                            },
+                            message: '"version" is required',
+                            path: ['version'],
+                            type: 'any.required'
+                        }
+                    ]);
+                }));
+
+        it('returns BAD REQUEST for template that cannot be parsed', () =>
+            server
+                .inject({
+                    method: 'POST',
+                    url: '/pipeline/template/validate',
+                    payload: {
+                        yaml: 'error: :'
+                    }
+                })
+                .then(reply => {
+                    assert.strictEqual(reply.statusCode, 400);
+
+                    const payload = JSON.parse(reply.payload);
+
+                    assert.match(payload.message, /YAMLException/);
+                }));
+
+        it('returns BAD REQUEST for invalid API input', () =>
+            server
+                .inject({
+                    method: 'POST',
+                    url: '/pipeline/template/validate',
+                    payload: { yaml: 1 }
+                })
+                .then(reply => {
+                    assert.strictEqual(reply.statusCode, 400);
+
+                    const payload = JSON.parse(reply.payload);
+
+                    assert.match(payload.message, /Invalid request payload input/);
+                }));
+    });
 });
